Manually run spectral leak as part of stage3 for MIRI
Spectral leak step requires 1B and 3A in the same ASN file, so at the end of stage3 we now manually run the spectral leak step with all channels in the same ASN. This ultimately means that   the 3A x1d file has the spectral leak step applied correctly by the end of stage3.

Including all channels in the ASN for the rest of stage3 can mess up the WCS pointing info for moving targets, so we process all channels together _only_ for the spectral leak step.